fix: address PR #956 review feedback and restore mock-agent compile
- Restore the closing brace for mockAvailableCommands() in
  apps/backend/cmd/mock-agent/main.go — its absence broke Backend
  (windows), Build, Run Backend Tests, Merge E2E Reports, and
  deploy-same-repo. Greptile flagged this as P0.
- Add hook unit tests for useResizableClarificationOverlay covering
  drag clamping (MIN_HEIGHT, 50% viewport cap), mouseup teardown,
  body cursor/userSelect lifecycle, ref-empty fallback, double-click
  reset, and unmount-mid-drag cleanup.
- Split clarification.spec.ts (755 → 599 lines) by extracting the
  /ask-single, /ask-multiple, and resize-layout describe blocks into
  apps/web/e2e/tests/chat/clarification-resize.spec.ts.
- Update stale "80vh" / "80%" comments in the hook JSDoc and resize
  spec to match the actual 50vh cap.
